pyproject.toml
[project]
name = "fastapi-sandbox"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "colorlog>=6.10.1", # 로그에 색 넣어줌
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.20",
    "asyncpg>=0.29.0",
    "alembic>=1.12.0",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0", # env 파일 불러올 수 있음
    "pydantic[email]>=2.4.0", # 타입힌트 활용해서 유효성 검사(@Valid), 직렬화 수행
    "redis>=5.0.0",
    "celery>=5.6.2",
    "httpx",
    "datetime",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
]
